Creates a study grid

This function is used to create a study grid for data integration pipelines and for data workflows. It is a simple wrapper around sf::st_make_grid.

pipegrid(x = NULL, bbox = NULL, cellsize, crs = 4326)

Arguments

x

object of class sf or sfc

bbox

bounding box to spatially subset the queried data, if applicable. The bounding box should be of the form c(xmin, ymin, xmax, ymax)

cellsize

target cellsize

crs

object of class crs; coordinate reference system of the target of the target grid in case argument x is missing, if x is not missing, its crs is inherited.

Value

soon described

Examples

if (FALSE) {
pipegrid()
}