Creates a new files from templates

This function is used to create new files from templates available in the pipedat package. This function is meant to facilitate writing of data pipeline for the pipedat developers and contributors. Templates available are data pipelines, data connects, and data workflows.

pipenew(name = NULL, template = "data_workflow")

Arguments

name

for a data pipeline or a data connect, name of the data that will be accessed through this new data pipeline. The name should be short and will only be used to ease the readability of the file structure, as each data pipeline is identified by a unique identifier rather than its name. For a data workflow, name of the data workflow; defaults to data_workflow

template

name of the template to generate, one of data_pipeline, data_connect, data_workflow. Defaults to data_workflow, as it is the most likely template to be needed by a user of the package.

Value

This function creates new files from the template available in inst/templates/

Examples

if (FALSE) {
# New data workflow
pipenew()

# New data pipeline
pipenew("dfo_logbooks", template = "data_pipeline")

# New data connect
pipenew("fisheries_intensity", template = "data_connect")
}